Move javascript out of lib/
So that users of this package can refer to its modules by, e.g.,
require('humbaba/runtime').

Also remove the symlink node_modules/humbaba, because it confuses npm
when it's trying to install this package as a dependency elsewhere.
As a result, our tests have to refer to humbaba modules by relative
paths.