lolo: Don't call functions with no arguments
Background: lolo_codegen treated all functions the same for simplicity,
including functions with no arguments.  This meant that users of those
"functions" had to call them explicitly.

Rationale 1: For example, consider this Lolo code:
    main = ioThen (putChar 'y') main
The AST for the body of main is
    ["ioThen", ["putChar", {"str": 'y'}], "main"]
But since main is a function with no arguments, under the old regime
it has to be
    ["ioThen", ["putChar", {"str": 'y'}], ["main"]]
lolo_parse can't know to do this because it doesn't know what the
identifier refers to (because it's a crude bootstrapping tool, not
a real compiler).

Rationale 2: It's annoying when writing Lolo-JSON by hand.